requirements.txt
dateparser==0.7.4
django-environ==0.4.5
pika==1.1.0
pip==20.1.1
psutil==5.7.0
psycopg2-binary==2.8.4
pymongo==3.9.0
PyVirtualDisplay==1.3.2
selenium==3.141.0
setuptools==41.6.0
requests==2.23.0
urllib3==1.25.9
wheel==0.33.6
xvfbwrapper==0.2.9
